add module to manage general Kea DHCP settings
Prior to this change, it was not possible to manage the general settings
of the Kea DHCP server as it can be done in the Web UI via
'Services->Kea DHCP->Settings'. I.e. it was not possible to configure
the interfaces on which the Kea should listen for DHCP requests.

This change adds a new module that allows to manage the general settings
of the Kea DHCP server. The code can be found in
'plugins/modules/dhcp_general.py' and
'plugins/module_utils/dhcp_general.py'